requirements.txt
pybit
python-dotenv
requests
flask
gunicorn